requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
opencv-python
nibabel
torch
torchvision
torchaudio
transformers
streamlit
jupyter
streamlit
torch
torchvision
Pillow
opencv-python


